Add additional metrics around server query performance
Track the number of running queries in the server as a gauge.
This will track all the queries received by netty workers which
is bound by the number of netty threads in simple request handler
but that can change with scheduled request handler. Moved logging
query timing numbers to request handler to have more timers and
better information on total query time.